requirements.txt
pandas
requests
numpy
beautifulsoup4
keras-tcn==2.1.2